\name{fICA-package}
\alias{fICA-package}
\docType{package}
\title{
\packageTitle{fICA}
}
\description{
\packageDescription{fICA}
}
\details{
\packageDESCRIPTION{fICA}
\packageIndices{fICA}
}

\references{
\cite{Miettinen, J., Nordhausen, K. and Taskinen, S. (2018), fICA: FastICA Algorithms and Their Improved Variants, \emph{The R Journal}, \bold{10}, 148--158, <doi:10.32614/RJ-2018-046>.}
}
\author{
\packageAuthor{fICA}

Maintainer: \packageMaintainer{fICA}

}
\keyword{package}
