\name{fICA-package}
\alias{fICA-package}
\docType{package}
\title{
Classical, Reloaded and Adaptive FastICA Algorithms
}
\description{
Algorithms for classical symmetric and deflation-based FastICA, reloaded deflation-based FastICA algorithm and an algorithm for adaptive deflation-based FastICA using multiple nonlinearities. For details, see Miettinen et al. (2014) <doi:10.1109/TSP.2014.2356442> and Miettinen et al. (2017) <doi:10.1016/j.sigpro.2016.08.028>.
 
}
\details{
\tabular{ll}{
Package: \tab fICA\cr
Type: \tab Package\cr
Version: \tab 1.1-0\cr
Date: \tab 2017-09-11\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Jari Miettinen <jari.p.miettinen@aalto.fi>

}
\keyword{package}
