\name{fICA-package}
\alias{fICA-package}
\docType{package}
\title{
Classical, Reloaded and Adaptive FastICA Algorithms
}
\description{
Algorithms for classical symmetric and deflation-based FastICA, reloaded deflation-based FastICA algorithm and an algorithm for adaptive deflation-based FastICA using multiple nonlinearities. 
}
\details{
\tabular{ll}{
Package: \tab fICA\cr
Type: \tab Package\cr
Version: \tab 1.0-3\cr
Date: \tab 2015-08-17\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Jari Miettinen <jari.p.miettinen@jyu.fi>

}
\keyword{package}
