\name{fICA-package}
\alias{fICA-package}
\docType{package}
\title{
Classical, Reloaded and Adaptive FastICA Algorithms
}
\description{
Algorithms for classical symmetric and deflation-based FastICA, reloaded deflation-based FastICA algorithm and an algorithm for adaptive deflation-based FastICA using multiple nonlinearities. For details, see Miettinen et al. (2014) <doi:10.1109/TSP.2014.2356442> and Miettinen et al. (2017) <doi:10.1016/j.sigpro.2016.08.028>.  The package is described in Miettinen, Nordhausen and Taskinen (2018) <doi:10.32614/RJ-2018-046>.
 
}
\details{
\tabular{ll}{
Package: \tab fICA\cr
Type: \tab Package\cr
Version: \tab 1.1-2\cr
Date: \tab 2021-12-08\cr
License: \tab GPL (>= 2)\cr
}

}

\references{
\cite{Miettinen, J., Nordhausen, K. and Taskinen, S. (2018), fICA: FastICA Algorithms and Their Improved Variants, \emph{The R Journal}, \bold{10}, 148--158, <doi:10.32614/RJ-2018-046>.}
}
\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Klaus Nordhausen <klaus.k.nordhausen@jyu.fi>

}
\keyword{package}
