\name{fICA-package}
\alias{fICA-package}
\docType{package}
\title{
Classic, reloaded and adaptive FastICA algorithms
}
\description{
The package includes classic symmetric and deflation-based FastICA algorithms, reloaded deflation-based FastICA and an algorithm for adaptive deflation-based FastICA using multiple contrast functions. 
}
\details{
\tabular{ll}{
Package: \tab fICA\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2013-11-24\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jari Miettinen, Klaus Nordhausen, Hannu Oja, Sara Taskinen

Maintainer: Jari Miettinen <jari.p.miettinen@jyu.fi>

}
\keyword{package}
